myrequirements.txt
absl-py==1.3.0
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==5.2.0
certifi==2020.12.5
charset-normalizer==2.1.1
click==8.1.3
cycler==0.10.0
docopt==0.6.2
flake8==3.9.1
flatbuffers==22.11.23
fuzzy-c-means==1.6.4
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.0.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.3.1
libclang==14.0.6
lxml==4.6.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.2
mccabe==0.6.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
Pillow==8.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.10.2
pyflakes==2.3.1
PyMaxflow==1.2.13
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
sh==1.14.2
six==1.15.0
soupsieve==2.2.1
tabulate==0.8.10
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm==4.61.0
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
yacs==0.1.8
zipp==3.11.0